+++ src/main/java/sayTheSpire/buffers/LeaderboardBuffer.java
@@ -1,5 +1,8 @@
 package sayTheSpire.buffers;
 
+import com.megacrit.cardcrawl.screens.leaderboards.LeaderboardEntry;
+import com.megacrit.cardcrawl.screens.stats.CharStat;
+
 public class LeaderboardBuffer extends Buffer {
 
     private int startIndex, endIndex;
@@ -8,6 +11,18 @@ public class LeaderboardBuffer extends Buffer {
         super("leaderboard", "leaderboard");
         this.startIndex = -1;
         this.endIndex = -1;
+    }
+
+    public void add(LeaderboardEntry entry) {
+        this.context.put("rank", entry.rank);
+        this.context.put("name", entry.name);
+        if (entry.isTime) {
+            this.context.put("score", CharStat.formatHMSM(entry.score));
+            this.addLocalized("content.timeEntry");
+        } else {
+            this.context.put("score", entry.score);
+            this.addLocalized("content.scoreEntry");
+        }
     }
 
     public void setIndices(int startIndex, int endIndex) {